Updated Msg for website
