+++ index.html
@@ -1,11 +1,11 @@
 <!DOCTYPE html>
 <html>
 <head>
-    <title>Menu</title>
+    <title>Calstate Long Beach Dining Hall Menu</title>
     <script src="https://cdnjs.cloudflare.com/ajax/libs/luxon/1.27.0/luxon.min.js"></script>
 </head>
 <body>
-    <h1>Menu</h1>
+    <h1>CSULB Dining Hall Menu(Only for testing, Menus displayed are not accurate)</h1>
       <h2>Beachside Menu</h2>
 
         <h3>Breakfast</h3>
@@ -31,6 +31,6 @@
         <h3>Dinner</h3>
           <div id="parkside_menu_dinner"></div>
 
-    <script src="chat.js"></script>
+    <script src="menu.js"></script>
 </body>
 </html>
